テスト用の共通処理をまとめた
